requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.0.0
matplotlib>=3.7.0
opencv-python>=4.7.0
tqdm>=4.65.0
faiss-cpu>=1.7.0
